+++ apex/__init__.py
@@ -1,1 +1,6 @@
+import os
+
 __version__ = '0.1.0'
+LOCAL_PATH = os.getcwd()
+
+
