+++ apex/core/calculator/Lammps.py
@@ -69,7 +69,7 @@ class Lammps(Task):
         if self.inter_type == "meam":
             model_lib = os.path.basename(self.model[0])
             model_file = os.path.basename(self.model[1])
-            os.chdir(os.path.join(output_dir, "../"))
+            os.chdir(os.path.join(output_dir, "../../"))
             if os.path.islink(model_lib):
                 link_lib = os.readlink(model_lib)
                 if not os.path.abspath(link_lib) == self.model[0]:
@@ -87,21 +87,21 @@ class Lammps(Task):
                 os.symlink(os.path.relpath(self.model[1]), model_file)
             os.chdir(output_dir)
             if not os.path.islink(model_lib):
-                os.symlink(os.path.join("..", model_lib), model_lib)
-            elif not os.path.join("..", model_lib) == os.readlink(model_lib):
+                os.symlink(os.path.join("../..", model_lib), model_lib)
+            elif not os.path.join("../..", model_lib) == os.readlink(model_lib):
                 os.remove(model_lib)
-                os.symlink(os.path.join("..", model_lib), model_lib)
+                os.symlink(os.path.join("../..", model_lib), model_lib)
 
             if not os.path.islink(model_file):
-                os.symlink(os.path.join("..", model_file), model_file)
-            elif not os.path.join("..", model_file) == os.readlink(model_file):
+                os.symlink(os.path.join("../..", model_file), model_file)
+            elif not os.path.join("../..", model_file) == os.readlink(model_file):
                 os.remove(model_file)
-                os.symlink(os.path.join("..", model_file), model_file)
+                os.symlink(os.path.join("../..", model_file), model_file)
             os.chdir(cwd)
 
         else:
             model_file = os.path.basename(self.model)
-            os.chdir(os.path.join(output_dir, "../"))
+            os.chdir(os.path.join(output_dir, "../../"))
             if os.path.islink(model_file):
                 link_file = os.readlink(model_file)
                 if not os.path.abspath(link_file) == self.model:
@@ -111,10 +111,10 @@ class Lammps(Task):
                 os.symlink(os.path.relpath(self.model), model_file)
             os.chdir(output_dir)
             if not os.path.islink(model_file):
-                os.symlink(os.path.join("..", model_file), model_file)
-            elif not os.path.join("..", model_file) == os.readlink(model_file):
+                os.symlink(os.path.join("../..", model_file), model_file)
+            elif not os.path.join("../..", model_file) == os.readlink(model_file):
                 os.remove(model_file)
-                os.symlink(os.path.join("..", model_file), model_file)
+                os.symlink(os.path.join("../..", model_file), model_file)
             os.chdir(cwd)
 
         dumpfn(self.inter, os.path.join(output_dir, "inter.json"), indent=4)
