+++ apex/core/calculator/lib/lammps.py
