+++ apex/core/property/EOS.py
@@ -8,7 +8,7 @@ from monty.serialization import dumpfn, loadfn
 
 import apex.calculator.lib.abacus as abacus
 import apex.calculator.lib.vasp as vasp
-import apex.calculator.lib.abacus_scf as abacus_scf
+import apex.core.calculator.lib.abacus_scf as abacus_scf
 from apex.core.Property import Property
 from apex.core.refine import make_refine
 from apex.core.reproduce import make_repro, post_repro
