--- apex/op/Collector.py
@@ -1,2 +1,0 @@
-
-
