+++ tests/test_abacus.py
@@ -8,11 +8,10 @@ from monty.serialization import loadfn
 
 from apex.calculator.ABACUS import ABACUS
 from apex.calculator.lib import abacus
-from apex.calculator.lib import abacus_scf
+from apex.core.calculator.lib import abacus_scf
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 __package__ = "tests"
-from .context import setUpModule
 
 
 class TestABACUS(unittest.TestCase):
