+++ tests/test_abacus_property.py
@@ -8,7 +8,7 @@ import numpy as np
 from monty.serialization import loadfn
 
 from apex.calculator.ABACUS import ABACUS
-from apex.calculator.lib import abacus_scf
+from apex.core.calculator.lib import abacus_scf
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 __package__ = "tests"
